+++ src/store/rooms.ts
@@ -26,10 +26,6 @@ type DroppedObject = {
 		x: number;
 		y: number;
 	};
-	group?: {
-		id: string;
-		groupIndex: number;
-	};
 	item: object;
 };
 
@@ -114,17 +110,11 @@ export class Rooms extends VuexModule {
 
 	@Action
 	async align(payload: DroppedObject): Promise<void> {
-		const { from, to, group } = payload;
+		const { from, to } = payload;
 		const reqObject = {
 			from,
 			to,
 		};
-
-		const url = group
-			? `/v3/dashboard/${this.gridElementsId}/moveElement/${group.id}/${group.groupIndex}`
-			: `/v3/dashboard/${this.gridElementsId}/moveElement`;
-
-		console.log("%crooms.ts line:127 url", "color: #007acc;", url); // TODO: remove
 
 		this.setLoading(true);
 		try {